+++ SwiggySpringBoot/swiggyClone/src/main/java/com/swiggy/swiggyClone/service/ApiService.java
@@ -276,7 +276,7 @@ public class ApiService {
 
 
     //Get user Address by id
-    public List<AddressTable> getUserAddressByUserID(Long id){
+    public List<AddressTable> getUserAddressByUserID(int id){
         return addressRepository.findByUserId(id);
 
     }
